Separate the core library into its own workspace
Some components of the library is also moved to the binary crate instead like the compilation env and its related traits. The parallel execution is also now handled with the Rayon library instead of implementing its own thread pool.
