+++ Cargo.toml
@@ -1,33 +1,14 @@
 [package]
-name = "texture-notes-v2"
-version = "2.0.0"
+name = "texture-notes-cli"
+version = "0.1.0"
 authors = ["foo-dogsquared <foo.dogsquared@gmail.com>"]
 edition = "2018"
-default-run = "texture-notes"
-cargo-features = ["named-profiles"]
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-chrono = { version = "0.4.10", features = ["serde"] }
 directories = "2.0.2"
-lazy_static = "1.4.0"
-globwalk = "0.7.1"
-handlebars = "2.0.2"
-heck = "0.3.1"
-regex = "1.3.1"
-serde = { version = "1.0", features = ["derive"] }
+rayon = "1.3.0"
 structopt = "0.3"
+texture-notes-lib = {path = "lib"}
 toml = "0.5.5"
-
-[dev-dependencies]
-tempfile = "3.1.0"
-
-[profile.release]
-opt-level = 'z'
-lto = true
-panic = 'abort'
-
-[[bin]]
-name = "texture-notes"
-path = "src/bin/main.rs"
