+++ lib/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "texture-notes-lib"
+version = "2.0.0"
+authors = ["foo-dogsquared <foo.dogsquared@gmail.com>"]
+edition = "2018"
+cargo-features = ["named-profiles"]
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+chrono = { version = "0.4.10", features = ["serde"] }
+lazy_static = "1.4.0"
+globwalk = "0.7.1"
+handlebars = "2.0.2"
+heck = "0.3.1"
+serde = { version = "1.0", features = ["derive"] }
+toml = "0.5.5"
+
+[dev-dependencies]
+tempfile = "3.1.0"
+
+[profile.release]
+opt-level = 'z'
+lto = true
+panic = 'abort'
