+++ lib/src/error.rs
@@ -1,4 +1,3 @@
-use std::convert::From;
 use std::error;
 use std::fmt;
 use std::io;
