+++ lib/src/masternote.rs
@@ -1,19 +1,15 @@
 use std::collections::HashMap;
 use std::fs::{self, OpenOptions};
 use std::path::PathBuf;
-use std::process;
-use std::str::FromStr;
 
 use toml;
 
 use crate::error::Error;
-use crate::helpers;
 use crate::note::Note;
-use crate::shelf::{self, Shelf, ShelfData, ShelfItem, ToCommand};
+use crate::shelf::{Shelf, ShelfData, ShelfItem};
 use crate::subjects::Subject;
 use crate::Object;
 use crate::Result;
-use crate::HANDLEBARS_REG;
 
 #[macro_use]
 use crate::modify_toml_table;
@@ -143,22 +139,5 @@ impl MasterNote {
     /// Return the file name of the master note.
     pub fn file_name(&self) -> String {
         MASTER_NOTE_FILE.to_string()
-    }
-}
-
-impl ToCommand for MasterNote {
-    fn to_command<S>(
-        &self,
-        cmd: S,
-    ) -> process::Command
-    where
-        S: AsRef<str>,
-    {
-        let cmd = cmd.as_ref();
-        let resulting_toml = format!("note = '{}'", self.file_name());
-        let note_as_toml = toml::Value::from_str(&resulting_toml).unwrap();
-        let command_string = HANDLEBARS_REG.render_template(&cmd, &note_as_toml).unwrap();
-
-        shelf::str_as_cmd(command_string)
     }
 }
