+++ lib/src/templates.rs
@@ -1,3 +1,8 @@
+//! An adapter for a template engine. 
+//! This is implemented in case Texture Notes decides to support multiple template engine.
+//! 
+//! (On the other hand, this may be just a case of overengineering.)
+
 use std::fs;
 use std::ops::Deref;
 use std::path::Path;
@@ -164,8 +169,7 @@ impl TemplateHandlebarsRegistry {
     }
 }
 
-/// A template is a Handlebars string to be rendered.
-/// This is specifically use in creating notes and other files that may need templating.
+/// A generic struct for templates to be used in a template engine.
 pub struct Template {
     name: String,
     s: String,
